LPS-89907 renaming package
