+++ modules/apps/data-engine/data-engine-api/bnd.bnd
@@ -6,6 +6,6 @@ Export-Package:\
 	com.liferay.data.engine.exception,\
 	com.liferay.data.engine.io,\
 	com.liferay.data.engine.model,\
-	com.liferay.data.engine.rules,\
+	com.liferay.data.engine.rule,\
 	com.liferay.data.engine.service,\
 	com.liferay.data.engine.storage
