+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/internal/rule/DEIntegerLiteralRuleFunction.java
@@ -12,31 +12,31 @@
  * details.
  */
 
-package com.liferay.data.engine.internal.rules;
+package com.liferay.data.engine.internal.rule;
 
 import com.liferay.data.engine.constants.DEDataDefinitionRuleConstants;
 import com.liferay.data.engine.model.DEDataDefinitionField;
-import com.liferay.data.engine.rules.DEDataDefinitionRuleFunction;
-import com.liferay.data.engine.rules.DEDataDefinitionRuleFunctionApplyRequest;
-import com.liferay.data.engine.rules.DEDataDefinitionRuleFunctionApplyResponse;
-import com.liferay.portal.kernel.util.Validator;
+import com.liferay.data.engine.rule.DEDataDefinitionRuleFunction;
+import com.liferay.data.engine.rule.DEDataDefinitionRuleFunctionApplyRequest;
+import com.liferay.data.engine.rule.DEDataDefinitionRuleFunctionApplyResponse;
 
-import java.util.stream.Stream;
+import org.apache.commons.lang.math.NumberUtils;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * It validates if a value is not empty.
+ * It validates if a value is an integer number.
  *
  * @author Leonardo Barros
  */
 @Component(
 	immediate = true,
-	property = {"de.data.definition.rule.function.name=" + DEDataDefinitionRuleConstants.EMPTY_RULE,
+	property = {"de.data.definition.rule.function.name=" + DEDataDefinitionRuleConstants.INTEGER_LITERAL_RULE,
 		"de.data.definition.rule.function.type=" + DEDataDefinitionRuleConstants.VALIDATION_RULE_TYPE},
 	service = DEDataDefinitionRuleFunction.class
 )
-public class DEEmptyRuleFunction implements DEDataDefinitionRuleFunction {
+public class DEIntegerLiteralRuleFunction
+	implements DEDataDefinitionRuleFunction {
 
 	@Override
 	/**
@@ -58,7 +58,7 @@ public class DEEmptyRuleFunction implements DEDataDefinitionRuleFunction {
 
 		deDataDefinitionRuleFunctionApplyResponse.setValid(false);
 		deDataDefinitionRuleFunctionApplyResponse.setErrorCode(
-			DEDataDefinitionRuleConstants.VALUE_MUST_NOT_BE_EMPTY_ERROR);
+			DEDataDefinitionRuleConstants.VALUE_MUST_BE_INTEGER_ERROR);
 
 		Object value = deDataDefinitionRuleFunctionApplyRequest.getValue();
 
@@ -66,19 +66,13 @@ public class DEEmptyRuleFunction implements DEDataDefinitionRuleFunction {
 			return deDataDefinitionRuleFunctionApplyResponse;
 		}
 
-		boolean result;
+		Integer valueInteger = NumberUtils.toInt(
+			value.toString(), Integer.MIN_VALUE);
 
-		if (isArray(value)) {
-			Object[] values = (Object[])value;
+		boolean result = false;
 
-			result = Stream.of(
-				values
-			).allMatch(
-				Validator::isNotNull
-			);
-		}
-		else {
-			result = Validator.isNotNull(value.toString());
+		if (valueInteger != Integer.MIN_VALUE) {
+			result = true;
 		}
 
 		deDataDefinitionRuleFunctionApplyResponse.setValid(result);
@@ -88,12 +82,6 @@ public class DEEmptyRuleFunction implements DEDataDefinitionRuleFunction {
 		}
 
 		return deDataDefinitionRuleFunctionApplyResponse;
-	}
-
-	protected boolean isArray(Object parameter) {
-		Class<?> clazz = parameter.getClass();
-
-		return clazz.isArray();
 	}
 
 }
